react-travel-app/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Flask==3.0.0
python-dotenv==1.0.0
Pillow>=9.0.0
PyJWT==2.8.1
python-jose[cryptography]==3.3.0
unidecode==1.3.7
Werkzeug==3.0.1
gunicorn==21.2.0
gunicorn==21.2.0
